Update Profile
I decided to separate profile and character create into two separate pages
