+++ profile.php
@@ -2,16 +2,16 @@
 <?php
     session_start();
     // Check that a user is logged in within the $_SESSION variable.
-
-?>
+    
+    ?>
 <html>
 
-    <head> 
-        <link rel="stylesheet" href="css/styles.css">
-    </head>
+<head>
+<link rel="stylesheet" href="css/styles.css">
+</head>
 
-<body>
-   
+<body style="background-color: grey;">
+
 <?php include('includes/header.php') ?>
 
 <div id = "row">
@@ -23,46 +23,7 @@
 <div id = "Characters">
 <?php getChars(); ?>
 </div>
-<div id = "CharCreation" style = "position: relative; bottom: 10px; display: none;">
-      <div id = "row" style = "position: relative; top: 20px;">
-        <div id="column">
-            Character Name: <input type="text" name="CharName" maxlength="20" size="20"><br>
-        </div>
-        <div id="column">
-            <select id="CharClass" style="position: relative; left: 1px;">
-                <option value="Select Class">Select Class</option>
-                <option value="Warrior">Warrior</option>
-                <option value="Mage">Mage</option>
-                <option value="Rogue">Rogue</option>
-                <option value="Necromancer">Necromancer</option>
-                <option value="Priest">Priest</option>
-            </select>
-           <select id="Weapon" style="position: relative; left: 5px;">
-                <option value="Select Weapon"> Select Weapon </option>
-                <?php getWeapons(); ?>
-            </select>
-            <select id="Armor" style="position: relative; left: 5px;">
-                <option value="Select Armor"> Select Armor </option>
-                <?php getArmors(); ?>
-            </select>
-            <button id="CharSub" style="position: relative; left: 5px;">Submit</button>
-        </div>   
-        </div>      
-</div>
-    
-<button class="btn" style="background-color: lightgrey; position: relative; left: 300px; top: 20px; bottom: 20px; border-radius: 0px;" onclick="expandCreation();"> + </button>
 
-
-<script>
-    function expandCreation(){
-        var toggle = document.getElementById("CharCreation");
-        if(toggle.style.display == "block"){
-            toggle.style.display = "none";
-        } else {
-            toggle.style.display = "block";
-        }
-    }
-</script> 
 </body>
 </html>
 
@@ -95,7 +56,7 @@
                 echo $toINSERT;
             }
         } else {
-            echo "0 results";
+            echo "No Characters to Show";
         }
         $conn->close();
     }
